pyproject.toml
[tool.poetry]
name = "sdss-archon"
version = "0.2.2-alpha.0"
description = "A library and actor to communicate with an STA Archon controller."
authors = ["José Sánchez-Gallego <gallegoj@uw.edu>"]
license = "BSD-3-Clause"
readme = "README.md"
homepage = "https://github.com/sdss/archon"
repository = "https://github.com/sdss/archon"
documentation = "https://sdss-archon.readthedocs.org"
keywords = ["astronomy", "software"]
classifiers = [
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Operating System :: OS Independent",
	"Programming Language :: Python :: 3.6",
	"Programming Language :: Python :: 3.7",
	"Programming Language :: Python :: 3.8",
    "Topic :: Documentation :: Sphinx",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
packages = [
    { include = "archon" }
]

[tool.poetry.scripts]
archon = "archon.__main__:archon"
lvm-lab = "archon.scripts.lvm_lab:lvm_lab"

[tool.poetry.dependencies]
python = "^3.8"
sdsstools = ">=0.4.0"
click = "^8.0.0"
numpy = "^1.19.5"
sdss-clu = "^1.0.0"
click-default-group = "^1.2.2"
daemonocle = "^1.1.1"
astropy = "^4.2"
sdss-drift = {version="^0.2.2", optional=true}

[tool.poetry.dev-dependencies]
ipython = ">=7.11.0"
flake8 = ">=3.7.9"
doc8 = ">=0.8.0"
pytest = ">=5.2.2"
pytest-asyncio = ">=0.10.0"
pytest-cov = ">=2.8.1"
pytest-mock = ">=1.13.0"
pytest-sugar = ">=0.9.2"
isort = ">=4.3.21"
codecov = ">=2.0.15"
coverage = {version = ">=5.0", extras = ["toml"]}
ipdb = ">=0.12.3"
rstcheck = ">=3.3.1"
Sphinx = ">=3.0.0"
black = ">=20.8b1"
jedi = "0.17.2"
sphinx-click = ">=2.6.0"
sphinx-jsonschema = ">=1.16.7"
myst-parser = ">=0.14.0"

[tool.poetry.extras]
lvm = ["sdss-drift"]

[tool.black]
line-length = 88
target-version = ['py39']
fast = true

[tool.isort]
profile = "black"
sections = ["FUTURE", "STDLIB", "TYPING", "THIRDPARTY", "SDSS", "FIRSTPARTY", "LOCALFOLDER"]
default_section = "THIRDPARTY"
known_first_party = "archon"
known_typing = ["typing"]
known_sdss = ["sdsstools"]
lines_after_imports = 2

[tool.pytest.ini_options]
addopts = "--cov archon --cov-report xml --cov-report html --cov-report term"
markers = [
    "commands: commands and replies for the test Archon"
]

[tool.coverage.run]
branch = true
include = ["archon/*"]
omit = [
    "*/__init__.py",
    "archon/__main__.py",
    "archon/actor/*",
    "archon/scripts/*",
    "archon/lvm/*",
    "archon/boss/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__"
]

[build-system]
requires = ["poetry-core>=1.0.0", "setuptools>=49.6.0"]
build-backend = "poetry.core.masonry.api"
